ejercicio terminado de carrera

+++ Integradores/CarreraRecreativa.java
@@ -1,7 +1,7 @@
 package Integradores;
 
 public class CarreraRecreativa extends Carrera {
-    public CarreraRecreativa(String nombre, double distancia) {
+    public CarreraRecreativa(String nombre, int distancia) {
         super(nombre, distancia);
     }
 }
